Update pipeline-ci-cd.yaml
Made some changes:
 
1. Check Running Containers:
This step checks if any container from the specified image is already running and exports its ID as an environment variable.

2. Stop and Remove Running Containers:
This step executes only if a container ID was found in the previous step. It stops and removes the container.

3. Deploy Docker Container:
This step deploys the new Docker container by running it on the specified port.